Update README.md
In my view there is no reason to limit the tracker to "scientific" conferences or?
Although it might start with academia I am sure it can / and will be used for all kind of 
conferences as the approach is exactly the same. In my view the whole effort is done to inspire all kinds of sectors to reduce carbon emissions related to conferences. Thus we should also reflect on the logo.